Refactor app into modular structure with db, scheduler, and routes
